[skip ci] wip: split GraphQL.Parser up into multiple modules
