+++ server/src-lib/Hasura/GraphQL/Parser/Internal/Parser.hs
@@ -1,0 +1,334 @@
+{-# OPTIONS_HADDOCK not-home #-}
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE StrictData          #-}
+
+-- | Defines the 'Parser' type and its primitive combinators.
+module Hasura.GraphQL.Parser.Internal.Parser where
+
+import           Hasura.Prelude
+
+import qualified Data.HashMap.Strict.Extended          as M
+import qualified Data.HashSet                          as S
+
+import           Control.Lens.Extended                 hiding (enum, index)
+import           Data.GADT.Compare.Extended
+import           Data.Int                              (Int32)
+import           Data.Parser.JSONPath
+import           Data.Type.Equality
+import           Data.Void
+import           Language.GraphQL.Draft.Syntax         (Description (..), EnumValue (..),
+                                                        Field (..), Name (..), Selection (..),
+                                                        SelectionSet, Value (..), litName, literal)
+
+import           Hasura.GraphQL.Parser.Internal.Class
+import           Hasura.GraphQL.Parser.Internal.Schema
+import           Hasura.Server.Utils                   (englishList)
+import           Hasura.SQL.Types
+
+-- -------------------------------------------------------------------------------------------------
+-- type definitions
+
+data Parser k m a = Parser
+  { pType   :: ~(Type k)
+  -- ^ Lazy for knot-tying reasons; see Note [Tying the knot] in
+  -- Hasura.GraphQL.Parser.Internal.Class.
+  , pParser :: ParserInput k -> m a
+  } deriving (Functor)
+
+instance HasName (Parser k m a) where
+  getName = getName . pType
+
+instance HasDefinition (Parser k m a) (TypeInfo k) where
+  definitionLens f parser = definitionLens f (pType parser) <&> \pType -> parser { pType }
+
+type family ParserInput k where
+  ParserInput 'Both = Value Variable
+  ParserInput 'Input = Value Variable
+  ParserInput 'Output = SelectionSet Variable
+
+-- | The constraint @(''Input' '<:' k)@ entails @('ParserInput' k ~ 'Value')@,
+-- but GHC can’t figure that out on its own, so we have to be explicit to give
+-- it a little help.
+inputParserInput :: forall k. 'Input <: k => ParserInput k :~: Value Variable
+inputParserInput = case subKind @'Input @k of { KRefl -> Refl; KBoth -> Refl }
+
+pInputParser :: forall k m a. 'Input <: k => Parser k m a -> Value Variable -> m a
+pInputParser = gcastWith (inputParserInput @k) pParser
+
+infixl 1 `bind`
+bind :: Monad m => Parser k m a -> (a -> m b) -> Parser k m b
+bind p f = p { pParser = pParser p >=> f }
+
+-- | Parses some collection of input fields. Build a 'FieldsParser' using
+-- 'field', 'fieldWithDefault', or 'fieldOptional', combine several together
+-- with the 'Applicative' instance, and finish it off using 'object' to turn it
+-- into a 'Parser'.
+data FieldsParser k m a = FieldsParser
+  -- Note: this is isomorphic to
+  -- `Compose ((,) [Definition (FieldInfo k)]) (ReaderT (HashMap Name (FieldInput k)) m) a`,
+  -- but working with that type kind of sucks.
+  { ifDefinitions :: [Definition (FieldInfo k)]
+  , ifParser      :: HashMap Name (FieldInput k) -> m a
+  } deriving (Functor)
+
+type family FieldInput k = r | r -> k where
+  FieldInput 'Input = Value Variable
+  FieldInput 'Output = Field Variable
+
+instance Applicative m => Applicative (FieldsParser k m) where
+  pure v = FieldsParser [] (const $ pure v)
+  a <*> b = FieldsParser
+    (ifDefinitions a <> ifDefinitions b)
+    (liftA2 (<*>) (ifParser a) (ifParser b))
+
+-- -------------------------------------------------------------------------------------------------
+-- combinators
+
+data ScalarRepresentation a where
+  SRBoolean :: ScalarRepresentation Bool
+  SRInt :: ScalarRepresentation Int32
+  SRFloat :: ScalarRepresentation Double
+  SRString :: ScalarRepresentation Text
+
+scalar
+  :: MonadParse m
+  => Name
+  -> Maybe Description
+  -> ScalarRepresentation a
+  -> Parser 'Both m a
+scalar name description representation = Parser
+  { pType = NonNullable $ TNamed $ mkDefinition name description TIScalar
+  , pParser = peelVariable >=> \v -> case representation of
+      SRBoolean -> case v of
+        VBoolean a -> pure a
+        _          -> typeMismatch name "a boolean" v
+      SRInt -> case v of
+        VInt a -> pure a
+        _      -> typeMismatch name "an integer" v
+      SRFloat -> case v of
+        VFloat a -> pure a
+        _        -> typeMismatch name "a float" v
+      SRString -> case v of
+        VString a -> pure a
+        _         -> typeMismatch name "a string" v
+  }
+
+boolean :: MonadParse m => Parser 'Both m Bool
+boolean = scalar $$(litName "Boolean") Nothing SRBoolean
+
+int :: MonadParse m => Parser 'Both m Int32
+int = scalar $$(litName "Int") Nothing SRInt
+
+float :: MonadParse m => Parser 'Both m Double
+float = scalar $$(litName "Float") Nothing SRFloat
+
+string :: MonadParse m => Parser 'Both m Text
+string = scalar $$(litName "String") Nothing SRString
+
+enum
+  :: MonadParse m
+  => Name
+  -> Maybe Description
+  -> NonEmpty (Definition EnumValueInfo, a)
+  -> Parser 'Both m a
+enum name description values = Parser
+  { pType = NonNullable $ TNamed $ mkDefinition name description $ TIEnum (fst <$> values)
+  , pParser = peelVariable >=> \case
+      VEnum (EnumValue value) -> case M.lookup value valuesMap of
+        Just result -> pure result
+        Nothing -> parseError $ "expected one of the values "
+          <> englishList "or" (dquoteTxt . dName . fst <$> values) <> "for type "
+          <> name <<> ", but found " <>> value
+      other -> typeMismatch name "an enum value" other
+  }
+  where
+    valuesMap = M.fromList $ over (traverse._1) dName $ toList values
+
+nullable :: forall k m a. (MonadParse m, 'Input <: k) => Parser k m a -> Parser k m (Maybe a)
+nullable parser = gcastWith (inputParserInput @k) Parser
+  { pType = case pType parser of
+      NonNullable t -> Nullable t
+      Nullable t    -> Nullable t
+  , pParser = peelVariable >=> \case
+      VNull -> pure Nothing
+      other -> Just <$> pParser parser other
+  }
+
+list :: forall k m a. (MonadParse m, 'Input <: k) => Parser k m a -> Parser k m [a]
+list parser = gcastWith (inputParserInput @k) Parser
+  { pType = NonNullable $ TList $ pType parser
+  , pParser = peelVariable >=> \case
+      VList values -> for (zip [0..] values) \(index, value) ->
+        withPath (Index index :) $ pParser parser value
+      other -> parseError $ "expected a list, but found " <> describeValue other
+  }
+
+object
+  :: MonadParse m
+  => Name
+  -> Maybe Description
+  -> FieldsParser 'Input m a
+  -> Parser 'Input m a
+object name description parser = Parser
+  { pType = NonNullable $ TNamed $ mkDefinition name description $
+      TIInputObject (ifDefinitions parser)
+  , pParser = peelVariable >=> \case
+      VObject fields -> do
+        -- check for extraneous fields here, since the FieldsParser just
+        -- handles parsing the fields it cares about
+        for_ (M.keys fields) \fieldName -> do
+          unless (fieldName `S.member` fieldNames) $
+            parseError $ name <<> " has no field named " <>> fieldName
+        ifParser parser fields
+      other -> typeMismatch name "an object" other
+  }
+  where
+    fieldNames = S.fromList (dName <$> ifDefinitions parser)
+
+field
+  :: (MonadParse m, 'Input <: k)
+  => Name
+  -> Maybe Description
+  -> Parser k m a
+  -> FieldsParser 'Input m a
+field name description parser = case pType parser of
+  NonNullable _ -> FieldsParser
+    { ifDefinitions = [mkDefinition name description case pType parser of
+        NonNullable typ -> IFRequired typ
+        Nullable typ    -> IFOptional typ (Just VNull)]
+    , ifParser = M.lookup name
+        >>> (`onNothing` parseError ("missing required field " <>> name))
+        >=> withPath (Key (unName name) :) . pInputParser parser
+    }
+  -- nullable fields just have an implicit default value of `null`
+  Nullable _ -> fieldWithDefault name description VNull parser
+
+fieldWithDefault
+  :: (MonadParse m, 'Input <: k)
+  => Name
+  -> Maybe Description
+  -> Value Void -- ^ default value
+  -> Parser k m a
+  -> FieldsParser 'Input m a
+fieldWithDefault name description defaultValue parser = FieldsParser
+  { ifDefinitions = [mkDefinition name description $
+      IFOptional (discardNullability $ pType parser) (Just defaultValue)]
+  , ifParser = M.lookup name >>> withPath (Key (unName name) :) . \case
+      Just value -> parseValue value
+      Nothing    -> pInputParser parser $ literal defaultValue
+  }
+  where
+    parseValue = \value -> case value of
+      VVariable (Variable { vValue, vDefinition }) ->
+        -- This case is tricky: if we get a nullable variable, we have to
+        -- pessimistically mark the query non-reusable, regardless of its
+        -- contents. Theoretically, we could be smarter: we could create a sort
+        -- of “derived variable” with its own default value and pass that to the
+        -- downstream parser. But that would be more complicated, so for now we
+        -- don’t do that.
+        -- FIXME: Clarify this explanation.
+        case dInfo vDefinition of
+          IFRequired _   -> parseValue value
+          IFOptional _ _ -> markNotReusable *> parseValue (literal vValue)
+      VNull -> pInputParser parser $ literal defaultValue
+      other -> pInputParser parser other
+
+-- | A nullable field with no default value. If the field is omitted, the
+-- provided parser /will not be called/. This allows a field to distinguish an
+-- omitted field from a field supplied with @null@ (which is permitted by the
+-- GraphQL specification). If you want a field that defaults to @null@, combine
+-- 'field' with 'nullable', instead.
+fieldOptional
+  :: (MonadParse m, 'Input <: k)
+  => Name
+  -> Maybe Description
+  -> Parser k m a
+  -> FieldsParser 'Input m (Maybe a)
+fieldOptional name description parser = FieldsParser
+  { ifDefinitions = [mkDefinition name description $
+      IFOptional (discardNullability $ pType parser) Nothing]
+  , ifParser = M.lookup name >>> withPath (Key (unName name) :) . traverse (pInputParser parser)
+  }
+
+selectionSet
+  :: MonadParse m
+  => Name
+  -> Maybe Description
+  -> FieldsParser 'Output m a
+  -> Parser 'Output m a
+selectionSet name description parser = Parser
+  { pType = NonNullable $ TNamed $ mkDefinition name description $ TIObject (ifDefinitions parser)
+  , pParser = \input -> do
+      let fields = input & mapMaybe \case
+            -- FIXME: handle fragments
+            SelectionField inputField -> Just inputField
+            _ -> Nothing
+      -- check for extraneous fields here, since the FieldsParser just
+      -- handles parsing the fields it cares about
+      for_ fields \Field{ _fName = fieldName } -> do
+        unless (fieldName `S.member` fieldNames) $
+          parseError $ name <<> " has no field named " <>> fieldName
+      ifParser parser $! M.fromListOn _fName fields
+  }
+  where
+    fieldNames = S.fromList (dName <$> ifDefinitions parser)
+
+type family SelectionResult k a b where
+  SelectionResult 'Both   a _ = (Name, a)
+  SelectionResult 'Output a b = (Name, a, b)
+
+selection
+  :: forall k m a b
+   . (MonadParse m, 'Output <: k)
+  => Name
+  -> Maybe Description
+  -> FieldsParser 'Input m a
+  -> Parser k m b
+  -> FieldsParser 'Output m (Maybe (SelectionResult k a b))
+selection name description argumentsParser bodyParser = FieldsParser
+  { ifDefinitions = [mkDefinition name description $
+      FieldInfo (ifDefinitions argumentsParser) (pType bodyParser)]
+  , ifParser = M.lookup name >>> withPath (Key (unName name) :) . traverse \selectionField -> do
+      let Field{ _fName = fieldName, _fArguments = arguments } = selectionField
+          alias = _fAlias selectionField & fromMaybe fieldName
+
+      for_ (M.keys arguments) \argumentName -> do
+        unless (argumentName `S.member` argumentNames) $
+          parseError $ name <<> " has no argument named " <>> argumentName
+      parsedArguments <- ifParser argumentsParser arguments
+
+      -- see Note [The delicate balance of GraphQL kinds] in Hasura.GraphQL.Parser.Internal.Schema
+      case subKind @'Output @k of
+        KRefl -> do
+          parsedBody <- pParser bodyParser $ _fSelectionSet selectionField
+          pure (alias, parsedArguments, parsedBody)
+        KBoth -> pure (alias, parsedArguments)
+  }
+  where
+    argumentNames = S.fromList (dName <$> ifDefinitions argumentsParser)
+
+-- -------------------------------------------------------------------------------------------------
+-- internal helpers
+
+peelVariable :: MonadParse m => Value Variable -> m (Value Variable)
+peelVariable (VVariable (Variable { vValue })) = markNotReusable $> literal vValue
+peelVariable value                             = pure value
+
+typeMismatch :: MonadParse m => Name -> Text -> Value Variable -> m a
+typeMismatch name expected given = parseError $
+  "expected " <> expected <> " for type " <> name <<> ", but found " <> describeValue given
+
+describeValue :: Value Variable -> Text
+describeValue = describeValueWith (describeValueWith absurd . vValue)
+
+describeValueWith :: (var -> Text) -> Value var -> Text
+describeValueWith describeVariable = \case
+  VVariable var -> describeVariable var
+  VInt _ -> "an integer"
+  VFloat _ -> "a float"
+  VString _ -> "a string"
+  VBoolean _ -> "a boolean"
+  VNull -> "null"
+  VEnum _ -> "an enum value"
+  VList _ -> "a list"
+  VObject _ -> "an object"
